+++ example/src/vite-env.d.ts
