+++ example/tsconfig.node.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../tsconfig.node.json"
+}
