+++ src/index.ts
@@ -1,0 +1,4 @@
+export { config } from "./wagmi";
+export { contractsByChain } from "./contracts";
+export type { ContractsByChain } from "./contracts";
+export { PoolzProvider } from "./PoolzProvider";
